+++ common-system-test/src/main/java/com/teamscale/test/commons/SystemTestUtils.java
@@ -18,7 +18,7 @@ public class SystemTestUtils {
 
 	/**
 	 * Turns the coverage of the given {@link TestInfo} into a string for simple assertions.
-	 *
+	 * <p>
 	 * Example: {@code file1.java:1,7-12;file2.java:9-22,33}
 	 */
 	public static String getCoverageString(TestInfo info) {