+++ common-system-test/src/main/java/com/teamscale/test/commons/TeamscaleMockServer.java
@@ -3,6 +3,7 @@ package com.teamscale.test.commons;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
+import com.teamscale.client.ClusteredTestDetails;
 import com.teamscale.client.PrioritizableTest;
 import com.teamscale.client.PrioritizableTestCluster;
 import com.teamscale.report.testwise.model.TestwiseCoverageReport;
@@ -36,6 +37,9 @@ public class TeamscaleMockServer {
 	private final JsonAdapter<List<PrioritizableTestCluster>> testClusterJsonAdapter = new Moshi.Builder().build()
 			.adapter(Types.newParameterizedType(List.class, PrioritizableTestCluster.class));
 
+	private final JsonAdapter<List<ClusteredTestDetails>> testDetailsJsonAdapter = new Moshi.Builder().build()
+			.adapter(Types.newParameterizedType(List.class, ClusteredTestDetails.class));
+
 	private final JsonAdapter<TestwiseCoverageReport> testwiseCoverageReportJsonAdapter = new Moshi.Builder().build()
 			.adapter(TestwiseCoverageReport.class);
 
@@ -43,6 +47,9 @@ public class TeamscaleMockServer {
 	public final List<String> uploadedReports = new ArrayList<>();
 	/** All user agents that were present in the received requests. */
 	public final Set<String> collectedUserAgents = new HashSet<>();
+
+	/** All tests that the test engine has signaled to Teamscale as being available for execution. */
+	public final Set<ClusteredTestDetails> availableTests = new HashSet<>();
 	private final Path tempDir = Files.createTempDirectory("TeamscaleMockServer");
 	private final Service service;
 	private final List<String> impactedTests;
@@ -73,8 +80,9 @@ public class TeamscaleMockServer {
 		return testwiseCoverageReportJsonAdapter.fromJson(uploadedReports.get(index));
 	}
 
-	private String handleImpactedTests(Request request, Response response) {
+	private String handleImpactedTests(Request request, Response response) throws IOException {
 		collectedUserAgents.add(request.headers("User-Agent"));
+		availableTests.addAll(testDetailsJsonAdapter.fromJson(request.body()));
 		List<PrioritizableTest> tests = impactedTests.stream().map(PrioritizableTest::new).collect(toList());
 		return testClusterJsonAdapter.toJson(Collections.singletonList(new PrioritizableTestCluster("cluster", tests)));
 	}